[#60] refactor : 업장 입력 수정 부분 리펙토링
