+++ catch-study/src/components/ManagementCafeInfo/ManagementCafeInfo.tsx
@@ -1,9 +1,4 @@
 import { useEffect, useState } from 'react';
-import { patchManagementInfo } from '../../apis/api/manager';
-import {
-  MANAGEMENT_INFO_ERROR,
-  ManagementErrorTypes,
-} from '../../config/error';
 import useManagementInfo from '../../hooks/useManagementInfo';
 import { FormDataTypes } from '../../types/management';
 import ManagementCafeInfoDisplay from '../ManagementCafeInfoDisplay/ManagementCafeInfoDisplay';
@@ -16,65 +11,9 @@ interface ManagementCafeInfoPropTypes {
 const ManagementCafeInfo: React.FC<ManagementCafeInfoPropTypes> = ({
   cafeInfo,
 }) => {
-  const {
-    roomInfos,
-    setRoomInfos,
-    usageFees,
-    setUsageFees,
-    formData,
-    setFormData,
-    handleInputChange,
-    handleNestedInputChange,
-    handleSelectChange,
-    handleRoomChange,
-    handleRoomNameChange,
-    handleAddRoom,
-    handleRemoveRoom,
-    handleAddFee,
-    handleRemoveFee,
-    handleFeeChange,
-    handleThumbnailChange,
-    handleStoreImagesChange,
-  } = useManagementInfo();
+  const { setRoomInfos, setUsageFees, setFormData } = useManagementInfo();
 
   const [isEdit, setIsEdit] = useState(false);
-
-  const getErrorMessage = (errorType: ManagementErrorTypes): string => {
-    return MANAGEMENT_INFO_ERROR[errorType];
-  };
-
-  const handleEditSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    console.log(formData);
-
-    let errorType: ManagementErrorTypes | null = null;
-
-    if (formData.cafe_name === '') {
-      errorType = 'CAFE_NAME_ERROR';
-    } else if (formData.seats === 0) {
-      errorType = 'SEATS_ERROR';
-    } else {
-      const timePattern = /^([01]\d|2[0-3]):([0-5]\d)$/;
-
-      if (!timePattern.test(formData.opening_hours)) {
-        errorType = 'OPENING_HOURS_ERROR';
-      } else if (!timePattern.test(formData.closed_hours)) {
-        errorType = 'CLOSED_HOURS_ERROR';
-      }
-    }
-
-    if (errorType !== null) {
-      alert(getErrorMessage(errorType));
-      return;
-    }
-
-    const rawData = await patchManagementInfo(formData);
-
-    // 올바른 요청일 때
-    setIsEdit(false);
-
-    console.log(rawData);
-  };
 
   useEffect(() => {
     setFormData(cafeInfo);
@@ -85,25 +24,7 @@ const ManagementCafeInfo: React.FC<ManagementCafeInfoPropTypes> = ({
   return (
     <>
       {isEdit ? (
-        <ManagementCafeInfoForm
-          formData={formData}
-          setFormData={setFormData}
-          handleEditSubmit={handleEditSubmit}
-          handleInputChange={handleInputChange}
-          handleNestedInputChange={handleNestedInputChange}
-          handleSelectChange={handleSelectChange}
-          usageFees={usageFees}
-          handleAddFee={handleAddFee}
-          handleRemoveFee={handleRemoveFee}
-          handleFeeChange={handleFeeChange}
-          roomInfos={roomInfos}
-          handleAddRoom={handleAddRoom}
-          handleRemoveRoom={handleRemoveRoom}
-          handleRoomChange={handleRoomChange}
-          handleRoomNameChange={handleRoomNameChange}
-          handleThumbnailChange={handleThumbnailChange}
-          handleStoreImagesChange={handleStoreImagesChange}
-        />
+        <ManagementCafeInfoForm setIsEdit={() => setIsEdit} />
       ) : (
         <ManagementCafeInfoDisplay
           cafeInfo={cafeInfo}
